requirements.txt
celery==4.1.1
Django==2.2.13
django-mongoengine==0.3.3
django-simple-menu==1.2.1
mock==2.0.0
mongoengine==0.18.2
